Remove Privacy Policy
The definitive text of the privacy policy for the Serval Mesh app is in
batphone/PRIVACY.md, beacuse it must be distributed with the app (in the
APK) and also in the app's source.

If we need a project-wide privacy policy (not oriented to any specific
app, but as an "umbrella" policy) then that can go in the Wiki and/or
the web site, and it should contain links to the app-specific policies
on GitHub for all the latest releases ('master' branches).